Updated insim to version 9 (LFS 0.6W+) * added packet size calculation, since LFS now states size divided by 4 * added CName parsing for modded vehicles (SkinID instead of 3 letter code)

+++ InSimDotNet/Packets/IPacket.cs
@@ -6,7 +6,7 @@
         /// <summary>
         /// Gets the size of the packet.
         /// </summary>
-        byte Size { get; }
+        int Size { get; }
 
         /// <summary>
         /// Gets the type of the packet.
